ref: Update targets on canceled runs too
